+++ src/components/Courses/tests/courses.test.js
@@ -1,18 +1,8 @@
 import '@testing-library/jest-dom';
-import {
-	screen,
-	fireEvent,
-	cleanup,
-	userEvent,
-	waitFor,
-} from '@testing-library/react';
-import Button from '../../../common/Button/Button';
-import { mockedState } from '../../../constants';
+import { fireEvent, waitFor } from '@testing-library/react';
+import { mockedState } from '../../../mockedContantsForTest';
 import { renderWithProviders } from '../../../testUtil';
-import CourseCard from '../components/CourseCard/CourseCard';
 import Courses from '../Courses';
-import { createMemoryHistory } from 'history';
-import { Navigate, useNavigate } from 'react-router-dom';
 
 const mockedUsedNavigate = jest.fn();
 
@@ -25,7 +15,7 @@ const renderComponent = () => {
 	return renderWithProviders(<Courses />);
 };
 
-describe('courses test', () => {
+describe('<Courses/> component test', () => {
 	test('courseCard equal to courses array length', () => {
 		const { getAllByTestId } = renderComponent();
 		let courseComp = getAllByTestId('courseCard');
