+++ src/components/Header/tests/header.test.js
@@ -1,28 +1,12 @@
 import React from 'react';
-import {
-	render,
-	cleanup,
-	fireEvent,
-	screen,
-	within,
-} from '@testing-library/react';
 import '@testing-library/jest-dom';
 import Header from '../Header';
-import { mockedState } from '../../../constants';
-import { Provider } from 'react-redux';
-import logoImage from '../../../../public/images/logo.png';
+import { mockedState } from '../../../mockedContantsForTest';
 import { renderWithProviders } from '../../../testUtil';
-// import { screen } from '@testing-library/dom';
-afterEach(cleanup);
-jest.useFakeTimers();
-const mockedStore = {
-	getState: () => mockedState,
-	subscribe: jest.fn(),
-	dispatch: jest.fn(),
-};
+
 const renderComponent = () => renderWithProviders(<Header />);
 
-describe('Header Tests', () => {
+describe('<Header/> component test cases', () => {
 	test('userName display', () => {
 		const { getByTestId } = renderComponent();
 		let userName = getByTestId('userName').innerHTML;
