+++ src/mockedContantsForTest.js
@@ -1,0 +1,33 @@
+export const mockedState = {
+	user: {
+		isAuth: true,
+		name: 'Test Name',
+		token: 'lkdjvnk',
+		role: 'admin',
+		email: 'test@email.com',
+		password: 'test123',
+	},
+	course: [
+		{
+			title: 'Ttitle',
+			description: 'Tdescription',
+			creationDate: '9/3/2021',
+			duration: 89,
+			authors: [
+				'9b87e8b8-6ba5-40fc-a439-c4e30a373d36',
+				'1c972c52-3198-4098-b6f7-799b45903199',
+			],
+			id: '66cc289e-6de9-49b2-9ca7-8b4f409d6467',
+		},
+	],
+	author: [
+		{ name: 'author', id: '9b87e8b8-6ba5-40fc-a439-c4e30a373d36' },
+		{ name: 'author2', id: '1c972c52-3198-4098-b6f7-799b45903199' },
+		{ name: 'author3', id: '072fe3fc-e751-4745-9af5-aa9eed0ea9ed' },
+		{ name: 'author4', id: '40b21bd5-cbae-4f33-b154-0252b1ae03a9' },
+		{ name: 'author5', id: '5e0b0f18-32c9-4933-b142-50459b47f09e' },
+		{ name: 'author6', id: '9987de6a-b475-484a-b885-622b8fb88bda' },
+		{ name: 'Sharguru', id: '23adc1df-8edc-4825-ab3a-2ffabf468806' },
+		{ name: 'lfjk', id: '8d64e193-b8e1-48d2-af3d-c87425343cc9' },
+	],
+};
