+++ src/testUtil.js
@@ -1,15 +1,13 @@
 import React from 'react';
 import { render } from '@testing-library/react';
 import { Provider } from 'react-redux';
-
 import { setupStore } from './store/index';
-import { mockedState } from './constants';
-import { BrowserRouter, Routes } from 'react-router-dom';
+import { mockedState } from './mockedContantsForTest';
+import { BrowserRouter } from 'react-router-dom';
 
 export function renderWithProviders(
 	ui,
 	{
-		// preloadedState = {},
 		preloadedState = mockedState,
 		// Automatically create a store instance if no store was passed in
 		store = setupStore(preloadedState),
